refactor(main): extract _register_phase4_services to backend/core/registrations/phase4 (audit A8, refs #153)
Commit 1 of N in PR A8 of the architectural-audit-2026-05-13 cycle (epic #145).

backend/main.py is 2572 lines with 53 service registrations. PR A8
extracts the registrations into per-domain modules under
backend/core/registrations/.

This commit extracts _register_phase4_services (~520 LOC). main.py
now has a 3-line shim that imports + delegates; lifespan() callers
need no changes.

  backend/core/registrations/__init__.py   (new)
  backend/core/registrations/phase4.py     (new, 562 LOC)
  backend/main.py                          (-509 LOC: 2572 -> 2063)

Behavior is bit-identical. The function body is moved verbatim except
for the public def name change to register(...).

Verification:
  poetry run python -c 'from backend.main import create_app; app = create_app()'
    -> OK
  poetry run pytest --no-cov -q
    -> 827 passed, 1 failed*, 28 skipped (matches main pre-extraction)
    *test_force_queue_processing pre-existing pollution.

Refs #145, refs #153
